// frontend/src/app/api/generated/model/ticketEventType.ts
/**
 * PulseDesk API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export enum TicketEventType {
  TICKET_CREATED = 'TICKET_CREATED',

  STATUS_CHANGED = 'STATUS_CHANGED',

  PRIORITY_CHANGED = 'PRIORITY_CHANGED',

  ASSIGNEE_CHANGED = 'ASSIGNEE_CHANGED',

  CATEGORY_CHANGED = 'CATEGORY_CHANGED',

  COMMENT_ADDED = 'COMMENT_ADDED',

  TICKET_RESOLVED = 'TICKET_RESOLVED',

  TICKET_CLOSED = 'TICKET_CLOSED',
}